+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ History
 
 * ``fix-format`` now only supports Python 3.6+.
 
+* ``fix-format`` can now run `black <https://github.com/python/black>`__ as the
+  code formatter for Python code.
+
+  See ``README.rst`` for instructions on how to use it.
+
+* By default ``fix-format`` is now less chatty, showing only files that were/would be changed and a summary
+  at the end, unless the new ``--verbose`` flag is given.
+
 1.8.0
 ----------